refactor: home and profile components
